Fix shift overflow on certain expressions
I got a report from libfuzz that some cryptic input could make the
assembler panic on shifts. It turns out that the check on whether the
operator was too big or not had to be explicitely casted to `usize` to
avoid signedness issues.

+++ lib/xixanta/src/assembler.rs
@@ -970,7 +970,7 @@ impl<'a> Assembler<'a> {
                 lval ^ rval
             }
             OperationType::Lshift => {
-                if rval > 16 {
+                if rval as usize > 16 {
                     return Err(Error {
                         line: node.value.line,
                         global: false,
@@ -983,7 +983,7 @@ impl<'a> Assembler<'a> {
                 lval << rval
             }
             OperationType::Rshift => {
-                if rval > 16 {
+                if rval as usize > 16 {
                     return Err(Error {
                         line: node.value.line,
                         global: false,
